+++ src/app/src/composables/useContext.ts
@@ -1,6 +1,7 @@
 import { createSharedComposable } from './createSharedComposable'
 import { computed, ref } from 'vue'
-import { StudioItemActionId, DraftStatus, StudioBranchActionId, StudioFeature,
+import {
+  StudioItemActionId, DraftStatus, StudioBranchActionId, StudioFeature,
 } from '../types'
 import type {
   PublishBranchParams,
@@ -23,6 +24,9 @@ import { useRoute, useRouter } from 'vue-router'
 import { findDescendantsFileItemsFromFsPath } from '../utils/tree'
 import { joinURL } from 'ufo'
 import { upperFirst } from 'scule'
+import { consola } from 'consola'
+
+const logger = consola.withTag('useContext')
 
 export const useContext = createSharedComposable((
   host: StudioHost,
@@ -113,10 +117,15 @@ export const useContext = createSharedComposable((
       const rootDocumentFsPath = joinURL(fsPath, 'index.md')
       const navigationDocumentFsPath = joinURL(fsPath, '.navigation.yml')
 
-      const navigationDocument = await host.document.db.create(navigationDocumentFsPath, `title: ${folderName}`)
-      const rootDocument = await host.document.db.create(rootDocumentFsPath, `# ${upperFirst(folderName)} root file`)
+      try {
+        const navigationDocument = await host.document.db.create(navigationDocumentFsPath, `title: ${folderName}`)
+        await activeTree.value.draft.create(navigationDocumentFsPath, navigationDocument)
+      }
+      catch (e) {
+        logger.warn(`Navigation document at path ${navigationDocumentFsPath} failed to create: ${e}`)
+      }
 
-      await activeTree.value.draft.create(navigationDocumentFsPath, navigationDocument)
+      const rootDocument = await host.document.db.create(rootDocumentFsPath, `# ${upperFirst(folderName)} root file`)
       const rootDocumentDraftItem = await activeTree.value.draft.create(rootDocumentFsPath, rootDocument)
 
       unsetActionInProgress()
@@ -221,9 +230,11 @@ export const useContext = createSharedComposable((
   const branchActionHandler: { [K in StudioBranchActionId]: (args: ActionHandlerParams[K]) => Promise<void> } = {
     [StudioBranchActionId.PublishBranch]: async (params: PublishBranchParams) => {
       const { commitMessage } = params
+      const prefix = host.meta.git?.commit?.messagePrefix
+      const finalMessage = prefix ? `${prefix} ${commitMessage.trim()}`.trim() : commitMessage.trim()
       const documentFiles = await documentTree.draft.listAsRawFiles()
       const mediaFiles = await mediaTree.draft.listAsRawFiles()
-      await gitProvider.api.commitFiles([...documentFiles, ...mediaFiles], commitMessage)
+      await gitProvider.api.commitFiles([...documentFiles, ...mediaFiles], finalMessage)
 
       // @ts-expect-error params is null
       await itemActionHandler[StudioItemActionId.RevertAllItems]()